Saber

+++ Saber/Saber.Infra/StringValueWrite.cs
@@ -223,12 +223,12 @@ public class StringValueWrite : Any
         long kk;
         kk = range.Count;
 
-        uint backSlash;
-        backSlash = (uint)stringComp.Char(classInfra.BackSlash, 0);
-        uint quote;
-        quote = (uint)stringComp.Char(classInfra.Quote, 0);
-        uint newLine;
-        newLine = (uint)stringComp.Char(classInfra.NewLine, 0);
+        long backSlash;
+        backSlash = stringComp.Char(classInfra.BackSlash, 0);
+        long quote;
+        quote = stringComp.Char(classInfra.Quote, 0);
+        long newLine;
+        newLine = stringComp.Char(classInfra.NewLine, 0);
         
         long countA;
         countA = 8;
